drivers/libhid.{c,h}: fix interrupt_size from int to size_t

+++ drivers/libhid.c
@@ -61,7 +61,7 @@ size_t max_report_size = 0;
 
 /* Tweaks for Powercom, at least */
 int interrupt_only = 0;
-int unsigned interrupt_size = 0;
+size_t interrupt_size = 0;
 
 /* ---------------------------------------------------------------------- */
 /* report buffering system */
@@ -494,7 +494,7 @@ int HIDGetEvents(hid_dev_handle_t udev, HIDData_t **event, int eventsize)
 	HIDData_t	*pData;
 
 	/* needs libusb-0.1.8 to work => use ifdef and autoconf */
-	buflen = comm_driver->get_interrupt(udev, buf, interrupt_size ? interrupt_size:sizeof(buf), 250);
+	buflen = comm_driver->get_interrupt(udev, buf, interrupt_size ? interrupt_size : sizeof(buf), 250);
 	if (buflen <= 0) {
 		return buflen;	/* propagate "error" or "no event" code */
 	}
